+++ lib/plugins/aws/package/compile/events/alexaSkill.js
