+++ lib/plugins/index.js
@@ -42,7 +42,7 @@ module.exports = [
   require('./aws/package/compile/events/stream/index.js'),
   require('./aws/package/compile/events/msk/index.js'),
   require('./aws/package/compile/events/alb/index.js'),
-  require('./aws/package/compile/events/alexaSkill/index.js'),
+  require('./aws/package/compile/events/alexaSkill.js'),
   require('./aws/package/compile/events/alexaSmartHome/index.js'),
   require('./aws/package/compile/events/iot/index.js'),
   require('./aws/package/compile/events/cloudWatchEvent/index.js'),
